+++ src/Controls/InflowMinEnvFlowControl.cpp
@@ -6,19 +6,19 @@
 
 InflowMinEnvFlowControl::InflowMinEnvFlowControl(
         int water_source_id, const vector<int> &aux_water_sources_ids,
-        const vector<double> *inflows, const vector<double> *releases)
+        const vector<float> *inflows, const vector<float> *releases)
         : MinEnvironFlowControl(water_source_id, aux_water_sources_ids,
                                 vector<int>(),
                                 INFLOW_CONTROLS),
           inflows(inflows),
           releases(releases) {}
 
-double InflowMinEnvFlowControl::getRelease(int week) {
-    double inflow =
+float InflowMinEnvFlowControl::getRelease(int week) {
+    float inflow =
             water_sources[water_source_id]->getUpstream_source_inflow() +
             water_sources[water_source_id]->getUpstreamCatchmentInflow();
 
-    double release = 0;
+    float release = 0;
     for (int i = 0; i < inflows->size(); ++i) {
         /// Done with ternary operator for improved performance.
         release = (inflow >= (*inflows)[i] ? (*releases)[i] : release);
