+++ src/DroughtMitigationInstruments/Transfers.h
@@ -13,34 +13,34 @@ class Transfers : public DroughtMitigationPolicy {
 private:
 
     const int source_utility_id;
-    const double source_treatment_buffer;
-    double average_pipe_capacity = 0;
+    const float source_treatment_buffer;
+    float average_pipe_capacity = 0;
     int transfer_water_source_id;
     map<int, int>* util_id_to_vertex_id = new map<int, int>();
     vector<int> buyers_ids;
-    vector<double> allocations;
-    vector<double> conveyed_volumes;
-    vector<double> buyers_transfer_triggers;
-    vector<double> flow_rates_and_allocations;
+    vector<float> allocations;
+    vector<float> conveyed_volumes;
+    vector<float> buyers_transfer_triggers;
+    vector<float> flow_rates_and_allocations;
     Utility *source_utility;
     WaterSource *transfer_water_source = nullptr;
-    Matrix<double> H, Aeq, A;
-    Vector<double> f, beq, b, allocations_aux, lb, ub;
+    Matrix<float> H, Aeq, A;
+    Vector<float> f, beq, b, allocations_aux, lb, ub;
 
 public:
 
     Transfers(
             const int id, const int source_utility_id,
-            int transfer_water_source_id, const double source_treatment_buffer,
+            int transfer_water_source_id, const float source_treatment_buffer,
             const vector<int> &buyers_ids,
-            const vector<double> &pipe_transfer_capacities,
-            const vector<double> &buyers_transfer_triggers,
-            const Graph utilities_graph, vector<double> conveyance_costs,
+            const vector<float> &pipe_transfer_capacities,
+            const vector<float> &buyers_transfer_triggers,
+            const Graph utilities_graph, vector<float> conveyance_costs,
             vector<int> pipe_owner);
 
     Transfers(const Transfers &transfers);
 
-    const vector<double> &getAllocations() const;
+    const vector<float> &getAllocations() const;
 
     ~Transfers();
 
@@ -50,8 +50,8 @@ public:
                                  vector<WaterSource *> water_sources,
                                  vector<MinEnvironFlowControl *> min_env_flow_controls) override;
 
-    vector<double>
-    solve_QP(vector<double> allocation_requests, double available_transfer_volume, double min_transfer_volume,
+    vector<float>
+    solve_QP(vector<float> allocation_requests, float available_transfer_volume, float min_transfer_volume,
                  int week);
 
     void setRealization(unsigned int realization_id) override;
